Add a way to filter window system events
This change introduces the class QAbstractWindowSystemEventFilter
which can be used to hook into QWindowSystemInterfacePrivate to
filter and get notified when window system events are being delivered.
One use case is to intercept key events from the underlying system
in QtCompositor and feed them into the xkbcommon state, and to modify
the events based on the resulting state.

Change-Id: I829eb7d960420135990fb0f6db54c14eea3e8e48